backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "amadeus>=12.0.0",
    "bcrypt==3.2.2",
    "cloudinary>=1.44.1",
    "confluent-kafka>=2.12.2",
    "email-validator>=2.2.0",
    "fastapi-guard>=4.2.2",
    "fastapi-mail>=1.5.0",
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.67.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pre-commit>=4.5.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "sqlmodel>=0.0.24",
    "typer>=0.16.0",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]
